+++ src/plonk/vanishing/prover.rs
@@ -9,6 +9,7 @@ use crate::{
     arithmetic::{eval_polynomial, CurveAffine, FieldExt},
     plonk::{ChallengeX, ChallengeY, Error},
     poly::{
+        self,
         commitment::{Blind, Params},
         multiopen::ProverQuery,
         Coeff, EvaluationDomain, ExtendedLagrangeCoeff, Polynomial,
@@ -62,19 +63,24 @@ impl<C: CurveAffine> Argument<C> {
 impl<C: CurveAffine> Committed<C> {
     pub(in crate::plonk) fn construct<
         E: EncodedChallenge<C>,
+        Ev: Copy + Send + Sync,
         R: RngCore,
         T: TranscriptWrite<C, E>,
     >(
         self,
         params: &Params<C>,
         domain: &EvaluationDomain<C::Scalar>,
-        expressions: impl Iterator<Item = Polynomial<C::Scalar, ExtendedLagrangeCoeff>>,
+        evaluator: poly::Evaluator<Ev, C::Scalar, ExtendedLagrangeCoeff>,
+        expressions: impl Iterator<Item = poly::Ast<Ev, C::Scalar, ExtendedLagrangeCoeff>>,
         y: ChallengeY<C>,
         mut rng: R,
         transcript: &mut T,
     ) -> Result<Constructed<C>, Error> {
         // Evaluate the h(X) polynomial's constraint system expressions for the constraints provided
-        let h_poly = expressions.fold(domain.empty_extended(), |h_poly, v| h_poly * *y + &v);
+        let h_poly = expressions
+            .reduce(|h_poly, v| &(&h_poly * *y) + &v) // Fold the gates together with the y challenge
+            .unwrap_or_else(|| poly::Ast::ConstantTerm(C::Scalar::zero()));
+        let h_poly = evaluator.evaluate(&h_poly, domain); // Evaluate the h(X) polynomial
 
         // Divide by t(X) = X^{params.n} - 1.
         let h_poly = domain.divide_by_vanishing_poly(h_poly);