+++ CHANGELOG.md
@@ -37,6 +37,11 @@ and this project adheres to Rust's notion of
 ### Removed
 - `halo2::arithmetic::BatchInvert` (use `ff::BatchInvert` instead).
 - `impl Default for halo2::poly::Rotation` (use `Rotation::cur()` instead).
+- `halo2::poly`:
+  - `EvaluationDomain::{add_extended, sub_extended, mul_extended}`
+  - `Polynomial::one_minus`
+  - `impl Neg, Sub for Polynomial`
+  - `impl Mul for Polynomial<_, ExtendedLagrangeCoeff>`
 
 ## [0.1.0-beta.1] - 2021-09-24
 Initial beta release!